[CORE-4] feat: Implement edit of user's fields route (#5)
**In authenticationController**
**_editUserAccount_**: Handles the user updates, validates password and
handles notifications for any changes made

**In authentciationMethods**
**_validateAndBuildUpdates_**: Validates and contructs the user update
object preparing the changes for DB update.
**_sendUpdateNotifications_**: Send Appropriate email notifications to
users based on what was changed(email, password or both) using the email
service.

**In authentication.utils**
**_verifyTokenAndGetUser_**: Abstracted the logic of verifying a token
and getting the user from it. It throws the appropriate errors for
invalid or expired tokens.

**In authentication.schema**
**_editUserSchema_**: Ensures optional email and password fields

**In authentication.types**
**_EditUserRequest_**: Defines the shape of an update request object for
user updates, used in _validateAndBuildUpdates_ to make it more
semantic.

**In routes**
Added the route to update the user's email and/or password, validating
the input data through the _editUserSchema_ and requiring a valid token
via the _verifyToken_ middleware

---------

Co-authored-by: mrvalraven <akatiago@gmail.com>
Co-authored-by: rafaelguerra <daRafael@users.noreply.github.com>